requirements.txt
streamlit>=1.31.0
langchain>=0.1.9
langchain-community>=0.0.24
langchain-openai>=0.0.6
chromadb>=0.4.22
pypdf>=4.0.1
sentence-transformers>=2.3.1
python-dotenv>=1.0.1
openai>=1.12.0
tiktoken>=0.6.0
requests>=2.31.0

